Reject actions for requests without permissions

+++ app/controllers/test_plans_controller.rb
@@ -15,37 +15,21 @@ class TestPlansController < ApplicationController
 
   # GET /projects/:project_id/test_plans
   def index
-    begin
-      @test_plans = TestPlan.where(project_id: @project.id).visible
-    rescue
-      flash.now[:error] = l(:error_project_not_found)
-      render 'forbidden', status: 404
-    end
+    @test_plans = TestPlan.where(project_id: @project.id).visible
   end
 
   # GET /projects/:project_id/test_plans/:id
   def show
-    begin
-      @test_case_test_plan = TestCaseTestPlan.new
-    rescue
-      flash.now[:error] = l(:error_test_plan_not_found)
-      render 'forbidden', status: 404
-    end
+    @test_case_test_plan = TestCaseTestPlan.new
   end
 
   # GET /projects/:project_id/test_plans/:id/edit
   def edit
-    begin
-      @test_plan = TestPlan.find(params.permit(:id)[:id])
-    rescue
-      flash.now[:error] = l(:error_test_plan_not_found)
-      render 'forbidden', status: 404
-    end
   end
 
   # PUT /projects/:project_id/test_plans/:id
   def update
-    @test_plan = TestPlan.find(params.permit(:id)[:id])
+    raise ::Unauthorized unless @test_plan.editable?
     update_params = {}
     update_params[:name] = test_plan_params[:name]
     update_params[:begin_date] = test_plan_params[:begin_date]
@@ -75,7 +59,9 @@ class TestPlansController < ApplicationController
 
   # POST /projects/:project_id/test_plans
   def create
-    find_project(params.permit(:project_id)[:project_id])
+    unless User.current.allowed_to?(:add_issues, @project, :global => true)
+      raise ::Unauthorized
+    end
     @test_plan = TestPlan.new(:name => test_plan_params[:name],
                               :begin_date => test_plan_params[:begin_date],
                               :end_date => test_plan_params[:end_date],
@@ -94,8 +80,9 @@ class TestPlansController < ApplicationController
 
   # DELETE /projects/:project_id/test_plans/:id
   def destroy
+    raise ActiveRecord::RecordNotFound.new unless @test_plan.visible?
+    raise ::Unauthorized unless @test_plan.deletable?
     begin
-      @test_plan = TestPlan.find(params.permit(:id)[:id])
       if @test_plan.destroy
         flash[:notice] = l(:notice_successful_delete)
         redirect_to project_test_plans_path
@@ -111,9 +98,11 @@ class TestPlansController < ApplicationController
 
   # POST /projects/:project_id/test_plans/:test_plan_id/assign_test_case
   def assign_test_case
-    authorize_with_issues_permission(params[:controller], :create)
+    return unless authorize_with_issues_permission(params[:controller], :create)
     begin
       @test_case = TestCase.find(params.require(:test_case_test_plan).permit(:test_case_id)[:test_case_id])
+      raise ActiveRecord::RecordNotFound.new unless @test_case.visible?
+      raise ActiveRecord::RecordNotFound.new unless @test_plan.visible?
       @test_case_test_plan = TestCaseTestPlan.where(test_plan: @test_plan,
                                                     test_case: @test_case).first
       unless @test_case_test_plan
@@ -131,8 +120,10 @@ class TestPlansController < ApplicationController
 
   # DELETE /projects/:project_id/test_plans/:test_plan_id/assign_test_case/:test_case_id
   def unassign_test_case
-    authorize_with_issues_permission(params[:controller], :destroy)
+    return unless authorize_with_issues_permission(params[:controller], :destroy)
     begin
+      raise ActiveRecord::RecordNotFound.new unless @test_case.visible?
+      raise ActiveRecord::RecordNotFound.new unless @test_plan.visible?
       @test_case_test_plan = TestCaseTestPlan.where(test_plan: @test_plan,
                                                     test_case: @test_case).first
       if @test_case_test_plan
